// src/app/models/administrator/entertainments.modal.ts
export interface entertainmentsModel {
  id: number;
  name: string;
  username: string;
  password: string;
  created_at: string;
  updated_at: string;
}

export interface sendDataAddModel {
  name: string;
  username: string;
  password: string;
}

export interface editDataModel {
  name: string;
  username: string;
  password: string;
}